updated the approveseat method
This function/method was changed to return an integer. This prevents an error that previously occured because the storePassengerName function was used even though a seat could be rejected. This cause a fatal error when trying to run the displayPassengerInfo method/function. Now that approve seat returns a value, an if statement was implemented in order to determine whether or not to call the storePassengerName function